postgres_table_to_csv/requirements.txt
sqlalchemy==1.4.26
psycopg2-binary==2.9.1
pandas==1.3.3
python-decouple==3.5